requirements.txt
kafka-python==2.0.2
psycopg2-binary==2.9.1
python-dotenv==0.19.1
requests==2.26.0
SQLAlchemy==1.4.25
urllib3==1.26.7